\name{pbat.help}
\alias{pbat.help}
\title{Help and Reporting Information}
\description{
  Provides some helpful suggestions and information if you are having
  some troubles getting the package set up (often tends to be version
  related). Information useful for reporting.
}
\usage{
pbat.help()
}
\references{
  \url{http://www.people.fas.harvard.edu/~tjhoffm/pbatR.html}  

  \url{http://www.biostat.harvard.edu/~clange/default.htm}
}
\seealso{
  \code{\link{pbat}}
}
\keyword{interface}
